Fix pickling SDMs
There was a bug whereby we were passing the features method
of the regressors into the SDM, which is not pickeable because
the method will be unbound when being serialized. I just
abstracted that into a callable to capture the state
and added a new method to return this new callable rather than
passing a reference to it.