Start moving the AAMBuilder into AAM
This moves the build function into a train function to match
the work I did on the SDM. Also, quite a big change is that
the scales are no longer flipped, and are now from lowest to
highest, again, to match SDM. This meant also changing some
fitting code. I need to check this more thoroughly to make
sure I'm doing it correctly. Similar changes need to be
done in the ATM/LK code.