Attempting to refactor SD-AAM
This attempts to refactor the Supervised Descent AAM code -
including attempting to get PO-CR implemented. It hasn't been
exhaustively tested, but I need to make a number of other changes
and I wanted a discrete commit for this refactoring work.

Changes:

The SupervisedDescentAAMFitter is a subclass of the
SupervisedDescentFitter. Therefore, it acts more like an SDM
than an AAM. This is great because it borrows a lot of code
from the SDFitter - which we are pretty happy with.

As an extension to this, the SDAAM algorithms have been totally
refactored. Now, they act exactly like SupervisedDescentAlgorithm,
but parametric. The major change is that the take an interface
directly rather than an interface class. Thus, the interface
properly encapsulates the AAM state rather than the algorithm.

Finally, accepting the refactoring of the train method to match
the SDAlgorithm class format (but with parameters), the interfaces
now don't have a circular dependancy on the SDAlgorithm.
The SDInterfaces now encapsulate warping and the transform. The
subclasses of SDAlgorithm, like ProjectOut are thus fairly simple.

The finaly thing is that results need to be refactored to now have
this dependancy on the whole algorithm. Therefore, I've added
a horrible shim that just properly returns the transform to
allow the algorithms to keep working for the time being.