Derive UnifiedAAMCLMFitter from MultiScaleParametricFitter
This is large change. Remove the builder module, the code is essentially moved in the base module in the UnifiedAAMCLM constructor scope. Using MultiScaleParametricFitter instead of the old Fitter base class.
The results are not conform. The scale transform is now computed by image.rescale, for scales [0.5, 1.0] the scaling between the first level and second level is ~2.09. In the previous version it was 2.0.
The initial error are also different. I will investigate this once I can quickly switch between commits.